Link to /blog with new tab
This is a hack because we don't have any kind of i18n in the Blog.
That's because we don't have the resources for translating every new
post.

In the future we would want to show a message in every i18n with an
alert box saying something like "You can translate this post to your
language" and a link to GitHub or Netlify CMS to enable readers to
contribute, but that's out of the scope at the moment.